Add update-answer: done create/update/delete/ answer logic

+++ src/components/common/QuestionForm/QuestionForm.jsx
@@ -11,6 +11,7 @@ import {
   addAnswerAction,
   addNewQuestionAction,
   deleteAnswerAction,
+  updateAnswerAction,
   updateQuestionAction,
 } from '@/store/features/test';
 
@@ -20,6 +21,8 @@ import Icon from '../Icon/Icon';
 import SortableItem from '../SortableItem/SortableItem';
 
 import s from './QuestionForm.module.scss';
+
+// TODO: добавить toaster для уведомлений валидации/создания/обновления вопросов и ответов
 
 const QuestionForm = ({ questionType, selectedQuestion, onCloseForm }) => {
   const [questionStep, setQuestionStep] = useState(1);
@@ -49,38 +52,63 @@ const QuestionForm = ({ questionType, selectedQuestion, onCloseForm }) => {
     if (selectedQuestion) {
       methods.setValue('title', selectedQuestion.title);
       methods.setValue('answers', selectedQuestion.answers);
-      update('answers', selectedQuestion.answers);
+      update(-1, selectedQuestion.answers);
+
+      const correctAnswer = selectedQuestion.answers.findIndex(
+        (a) => a.is_right
+      );
+      setCorrectAnswer(correctAnswer);
     }
   }, [selectedQuestion]);
 
   const handleSubmit = methods.handleSubmit((data) => {
     if (selectedQuestion) {
-      dispatch(
-        updateQuestionAction({
-          questionId: selectedQuestion.id,
-          title: data.title,
-          question_type: selectedQuestion.question_type,
-        })
-      );
+      const oldTitle = selectedQuestion.title;
+      const oldAnswers = selectedQuestion.answers;
+
+      const isAnswersChanged =
+        JSON.stringify(oldAnswers) !== JSON.stringify(data.answers);
+
+      if (oldTitle !== data.title) {
+        dispatch(
+          updateQuestionAction({
+            questionId: selectedQuestion.id,
+            title: data.title,
+            question_type: selectedQuestion.question_type,
+          })
+        );
+      }
+
+      // Определяю, есть ли изменения в ответах и патчу ответы
+      if (isAnswersChanged) {
+        data.answers.forEach((_, index) => {
+          const newAnswer = data.answers[index];
+          const oldAnswer = oldAnswers[index];
+          if (
+            oldAnswer.text !== newAnswer.text ||
+            oldAnswer.is_right !== newAnswer.is_right
+          ) {
+            dispatch(
+              updateAnswerAction({
+                answerId: oldAnswer.id,
+                questionId: selectedQuestion.id,
+                newData: newAnswer,
+              })
+            );
+          }
+        });
+      }
       return;
     }
 
-    // TODO: Validation
-    if (questionType === 'single') {
+    if (questionType === 'single' || questionType === 'multiple') {
       const title = data.title;
-      const answers = data.answers.map((answer, idx) => {
-        return {
-          text: answer.text,
-          is_right: correctAnswer === fields[idx].id,
-        };
-      });
 
       dispatch(
         addNewQuestionAction({
           title,
           question_type: questionType,
           answer: 1,
-          answers,
           testId: test.id,
         })
       );
@@ -88,19 +116,14 @@ const QuestionForm = ({ questionType, selectedQuestion, onCloseForm }) => {
       setQuestionStep(2);
     }
 
-    if (questionType === 'multiple') console.log(data); // TODO: add question with type "multiple" (api call)
-
     if (questionType === 'number') console.log(data); // TODO: add question with type "number" (api call)
   });
 
-  const handleChangeCorrectAnswer = (id) => {
-    setCorrectAnswer(id);
-  };
-
   const handleAddAnswer = () => {
-    if (questionType !== 'number' || fields.length < 1) {
+    if (questionType !== 'number') {
       append({ text: '', is_right: false });
     }
+
     setIsAnswerEditing(true);
   };
 
@@ -108,6 +131,11 @@ const QuestionForm = ({ questionType, selectedQuestion, onCloseForm }) => {
     const questionId = test.questions[test.questions.length - 1].id;
     const allAnswers = methods.getValues().answers;
     const newAnswer = allAnswers[allAnswers.length - 1];
+
+    if (newAnswer.text === '') {
+      alert('Введите текст ответа');
+      return;
+    }
 
     dispatch(
       addAnswerAction({
@@ -119,15 +147,24 @@ const QuestionForm = ({ questionType, selectedQuestion, onCloseForm }) => {
   };
 
   const handleDeleteAnswer = (id) => {
+    if (methods.getValues().answers.length < 3) {
+      alert(
+        'Нельзя удалить вопрос, если количество вариантов ответа не менее 3'
+      );
+      return;
+    }
     remove(id);
-    const answerId = selectedQuestion.answers[id].id;
 
-    dispatch(
-      deleteAnswerAction({
-        questionId: selectedQuestion.id,
-        answerId,
-      })
-    );
+    if (selectedQuestion) {
+      const answerId = selectedQuestion.answers[id].id;
+
+      dispatch(
+        deleteAnswerAction({
+          questionId: selectedQuestion.id,
+          answerId,
+        })
+      );
+    }
   };
   // TODO: add changing position field of answer in API
   const handleDragEnd = (event) => {
@@ -142,11 +179,42 @@ const QuestionForm = ({ questionType, selectedQuestion, onCloseForm }) => {
     }
   };
 
+  const handleValidate = () => {
+    if (questionStep === 1) {
+      return true;
+    }
+
+    if (questionType === 'single' || questionType === 'multiple') {
+      if (fields.length < 2) {
+        alert('Добавьте хотя бы два варианта ответа');
+        return false;
+      }
+
+      if (correctAnswer === null) {
+        alert('Выберите правильный ответ');
+        return false;
+      }
+
+      return true;
+    }
+    // TODO: add validation for "number"
+  };
+
   const handleCloseForm = () => {
-    setQuestionStep(1);
-    setCorrectAnswer(null);
-    methods.reset();
-    onCloseForm();
+    if (handleValidate()) {
+      setQuestionStep(1);
+      setCorrectAnswer(null);
+      methods.reset();
+      onCloseForm();
+    }
+  };
+
+  const handleChangeCorrectAnswer = (field, index) => {
+    const oldCorrectAnswerIdx = fields.findIndex((f) => f.is_right === true);
+
+    setCorrectAnswer(index);
+    methods.setValue(`answers.${index}.is_right`, true);
+    methods.setValue(`answers.${oldCorrectAnswerIdx}.is_right`, false);
   };
 
   return (
@@ -202,10 +270,16 @@ const QuestionForm = ({ questionType, selectedQuestion, onCloseForm }) => {
                           className={s.checkbox}
                           type='checkbox'
                           fieldName={`answers.${index}.is_right`}
-                          checked={correctAnswer === field.id}
-                          onClick={() => handleChangeCorrectAnswer(field.id)}
+                          checked={correctAnswer === index}
+                          onClick={() =>
+                            handleChangeCorrectAnswer(field, index)
+                          }
                         />
-                        <div onClick={() => handleDeleteAnswer(index)}>
+                        <div
+                          onClick={() => {
+                            handleDeleteAnswer(index);
+                          }}
+                        >
                           <Icon name='delete' className={s.delete} />
                         </div>
                       </div>
